feat: Implement loading state with SkeletonLoader; remove BigSpinner; enhance AuthContext with profile image handling
