+++ src/App.tsx
@@ -1,15 +1,23 @@
 import { RouterProvider, createBrowserRouter } from 'react-router-dom';
 import { Home, Login, NotFound, About, Products as MainProducts, ProductDetails as MainProductDetails, Contact, Profile } from './pages';
-import { Header, BigSpinner, PrivateRoute, HeaderAdmin } from './components';
+import { Header, PrivateRoute, HeaderAdmin, SkeletonLoader } from './components';
 import { Customers, Dashboard, Inventory, Orders, ProductDetails, Products } from './pages/admin'
 import { useEffect, useState } from 'react';
 import { SignUp } from './pages/SignUp';
-import { AuthProvider } from './context';
+import { AuthProvider, useAuth } from './context';
 import { account } from './appwrite';
 
 const App = () => {
+  return (
+    <AuthProvider>
+      <AppContent />
+    </AuthProvider>
+  );
+};
 
+const AppContent = () => {
   const [theme, setTheme] = useState('dark');
+  const { loading } = useAuth();
 
   useEffect(() => {
     const fetchTheme = async () => {
@@ -27,7 +35,9 @@ const App = () => {
 
   const toggleTheme = async () => {
     setTheme(theme === 'dark' ? 'light' : 'dark');
-    await account.updatePrefs({ theme: theme === 'dark' ? 'light' : 'dark' });
+    const prefs = await account.getPrefs();
+    prefs.theme = theme === 'dark' ? 'light' : 'dark';
+    await account.updatePrefs({ ...prefs });
   };
 
   const router = createBrowserRouter([
@@ -115,11 +125,13 @@ const App = () => {
   );
 
   return (
-    <AuthProvider>
-      <div className={`${theme === 'dark' ? 'bg-neutral-950 text-white' : 'bg-white text-neutral-950'}`}>
-        <RouterProvider router={router} fallbackElement={<BigSpinner />} />
-      </div>
-    </AuthProvider>
+    <div className={`${theme === 'dark' ? 'bg-neutral-950 text-white' : 'bg-white text-neutral-950'}`}>
+      {
+        loading ?
+          <SkeletonLoader /> :
+          <RouterProvider router={router} fallbackElement={<SkeletonLoader />} />
+      }
+    </div>
   );
 };
