--- src/components/BigSpinner.tsx
@@ -1,8 +1,0 @@
-
-export const BigSpinner = () => {
-    return (
-        <div className="flex items-center justify-center h-screen">
-            <div className="animate-spin rounded-full h-32 w-32 border-t-4 border-b-4 border-gold"></div>
-        </div>
-    );
-};
