+++ src/components/index.ts
@@ -1,5 +1,4 @@
 export * from "./Header";
-export * from "./BigSpinner";
 export * from "./PrivateRoute";
 export * from "./SkeletonLoader";
 export * from "./CartPopUp";
